moved useEnvironment to eliminate hooks folder
